Change asset compilation system
The previous system required everything to be asynchronous, but not everything had to be.
This meant that if too many assets were being loaded, sometimes compilation tasks would be stuck in the queue behind threads blocking on them: deadlock.
Also somethings were forced to be asynchronous or be weird that shouldn't have been.
So yeah this is probably better.