Updated immediate system to create objects on the fly
The previous method of detecting threads and creating immediates ahead of time had a tendency to fail to find all the t>
The new system creates them on the fly if they haven't been previously created (and has a nicer API now).

The current version does not yet work on the device I tested on, it's having device loss issues. Now I have to hunt those down (yay)...