Begin shader and material system rewrite
The whole shader and material system was full of bad idea juice, so it's time to fix it. I'm starting with Vulkan because that's what made me do this. OpenGL is second. Also yes, the Vulkan backend is in a broken state now. No I don't care. The dev branch isn't always stable.